Make JMX setup more generic.
Because the JMX setup needs to be applied to every component running in its
own JVM (which includes the Tomcat core.war and the Heritrix3 agent), move the
JMX setup into its own section which is referenced in multiple places.

Added link to the top of quick-start-guide.rst, release-notes.rst,
upgrade-guide.rst and user-manual.rst so they can be referenced from other
documents.